+++ tests/test_cases/predator_prey_model.py
@@ -1,0 +1,7 @@
+from tests.test_cases.predator_prey_model_source.state_update_blocks import state_update_blocks
+from tests.test_cases.predator_prey_model_source.state_variables import initial_state
+from tests.test_cases.predator_prey_model_source.sys_params import params
+from tests.test_cases.predator_prey_model_source import run
+
+MONTE_CARLO_RUNS = 3
+TIMESTEPS = 300